requirements.txt
cassandra-driver
PrettyTable

